Restructure project to follow best practice
