--- api/handler/handler.go
@@ -1,6 +1,0 @@
-package handler
-
-type IHandler[T any] interface {
-	Get() T
-	GetById(id int) T
-}
