+++ api/internal/api/handler/handler.go
@@ -1,0 +1,6 @@
+package handler
+
+type IHandler[T any] interface {
+	Get() T
+	GetById(id int) T
+}
