+++ api/internal/model/image.go
@@ -1,0 +1,5 @@
+package model
+
+type Image struct {
+	Full string `json:"full"`
+}
