--- api/model/image.go
@@ -1,5 +1,0 @@
-package model
-
-type Image struct {
-	Full string `json:"full"`
-}
